requirements.txt
anyio==4.11.0
attrs==25.4.0
Authlib==1.6.5
avro==1.12.1
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
confluent-kafka==2.12.2
cryptography==46.0.3
elastic-transport==9.2.0
elasticsearch==9.2.0
fastavro==1.12.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
pycparser==2.23
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
requests==2.32.5
six==1.17.0
sniffio==1.3.1
typing_extensions==4.15.0
urllib3==2.5.0
